add 4.21*: Blog List Expansion, step 9

+++ part4/controllers/blog.js
@@ -36,6 +36,18 @@ blogRouter.post('/', async (request, response) => {
 
 blogRouter.delete('/:id', async (request, response) => {
     const { id } = request.params
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(id)
+
+    if (user._id.toString() !== blog.user.toString()) {
+        return response.status(401).json({ error: 'You cannot delete someone elses blog' })
+    }
+
     await Blog.findByIdAndDelete(id)
     return response.status(204).end()
 })
